move report filter factory to filters module
